% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{feedback_get_formatted_string}
\alias{feedback_get_formatted_string}
\title{Format the feedback string}
\usage{
feedback_get_formatted_string(print_this, type, findme, prefix, suffix)
}
\value{
Returns a properly an consistent formatted string containing
  the parameters handed over to this function.
}
\description{
Helper function for the feedback function to combine the input
  parameters in proper manner to ge a pretty and informative string which
  than can be added to the logfile and/or be displayed in the console.
  CAUTION: 'print_this' must be of length 1! For arrays loop through them
  by hand and call this function several times!
  Internal use. Use the robust 'feedback' function instead.
}
