% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{feedback_to_console}
\alias{feedback_to_console}
\title{Print to the console. Internal use only.}
\usage{
feedback_to_console(
  print_this,
  type,
  findme,
  prefix,
  suffix,
  logjs,
  logfile_dir,
  headless = TRUE
)
}
\value{
No return value, called for side effects (see description)
}
\description{
Helper function for the feedback function to print
  stuff to the console. Everything will also be added to the logfile.
  Internal use. Use the robust 'feedback' function instead.
}
