% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{feedback_to_logfile}
\alias{feedback_to_logfile}
\title{Add to the logfile. Internal use.}
\usage{
feedback_to_logfile(print_this, type, findme, prefix, suffix, logfile_dir)
}
\value{
No return value, called for side effects (see description)
}
\description{
Helper function for the feedback function to add content
  to the logfile. Internal use.
  Use the robust 'feedback' function instead.
}
