% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_lines_to_list.R
\name{file_lines_to_list}
\alias{file_lines_to_list}
\title{Read in lines from a file and store it in a list.}
\usage{
file_lines_to_list(filepath)
}
\arguments{
\item{filepath}{(string) Path to file to read in.}
}
\value{
A list with one element per row of the input file
}
\description{
Read in lines from a file and store it in a list.
}
