% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{log_get_current_options}
\alias{log_get_current_options}
\title{Get the current settings for the logging function as list.}
\usage{
log_get_current_options()
}
\value{
The list with the current parameters.
}
\description{
Get the current settings for the logging function as list
}
\examples{
  log_get_current_options()

}
