% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{log_get_default_options}
\alias{log_get_default_options}
\title{Get the default settings for the logging function as list.}
\usage{
log_get_default_options()
}
\value{
The list with the default parameters.
}
\description{
Get the default settings for the logging function as list
}
\examples{
  log_get_default_options()

}
