% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{log_map_type_to_loggertype}
\alias{log_map_type_to_loggertype}
\title{Get the logger type from the type string (the argument of the
  `feedback()` function)}
\usage{
log_map_type_to_loggertype(type)
}
\value{
The `logger`type. If no corresponding logger-type is found,
  the result will be `NULL`.
}
\description{
Mapping the log-types from string to logger::<type>.
}
\examples{
  log_map_type_to_loggertype("error")
  ## will return: `logger::ERROR`

}
