% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{log_remove_options}
\alias{log_remove_options}
\title{Remove all log-related options from `options()`.}
\usage{
log_remove_options()
}
\value{
Nothing.
}
\description{
Remove all log-related options from `options()`.
}
\examples{
  log_remove_options()

}
