% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log_set_defaults}
\alias{log_set_defaults}
\title{Set default options for all log-functions}
\usage{
log_set_defaults(
  print_this = NULL,
  type = NULL,
  ui = NULL,
  console = NULL,
  logfile = NULL,
  logjs = NULL,
  prefix = NULL,
  suffix = NULL,
  findme = NULL,
  logfile_dir = NULL,
  headless = NULL
)
}
\value{
No return value, called for side effects (see description)
}
\description{
This function sets the default log options. Parameters not
  supplied to this function will be set with the default value.
}
\examples{
  DIZtools::set_log_options()

}
