% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_to_position.R
\name{number_to_position}
\alias{number_to_position}
\title{Converts an integer number to its "verbal position".
  1 --> "1st", 2 --> "2nd", 3 --> "3rd", 4 --> "4th", ...}
\usage{
number_to_position(x)
}
\arguments{
\item{x}{A number.}
}
\value{
Returns the input number as string with a new suffix depending on
 the numbers position in the numbers bar.
}
\description{
Converts an integer number to its "verbal position".
  1 --> "1st", 2 --> "2nd", 3 --> "3rd", 4 --> "4th", ...
}
\examples{
{

}
}
