% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setenv_file.R
\name{setenv_file}
\alias{setenv_file}
\title{Set all variables of a `.env` file to the system environment.}
\usage{
setenv_file(env_file)
}
\arguments{
\item{env_file}{A character. The full path including the file name to
the file containing the environment variable definitions to be loaded.}
}
\value{
No return value, called for side effects (see description)
}
\description{
Internal function to set environment variables that are
  necessary for the database connections with \code{db_connection}.
  Old function name: `set_env_vars()`.
}
\examples{
\dontrun{set_env_vars("./.env")}

}
\seealso{
\code{Sys.setenv}
}
