% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_diff_print.R
\name{time_diff_print}
\alias{time_diff_print}
\title{Calculate time difference of two timestamps, round the value and
  return a string with suiting unit.}
\usage{
time_diff_print(
  older_timestamp,
  newer_timestamp,
  digits = 2,
  thousands_separator = "",
  decimal_separator = "."
)
}
\arguments{
\item{older_timestamp}{(POSIXct) Start time.}

\item{newer_timestamp}{(POSIXct) End time.}

\item{digits}{(int, optional) The number of digits after the decimal
separator to round to.}

\item{thousands_separator}{(string, optional) Used as mark between every 3
decimals before the decimal point.}

\item{decimal_separator}{(string, optional) The character to be used to
indicate the numeric decimal point.}
}
\value{
A list with one element per row of the input file
}
\description{
Create string with time difference in suitable unit.
}
