% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{log_map_type_to_loggertype}
\alias{log_map_type_to_loggertype}
\title{Get the logger type from the type string (the argument of the
  `feedback()` function)}
\usage{
log_map_type_to_loggertype(type)
}
\arguments{
\item{type}{(String) The type of the message. E.g. "error", "Info".}
}
\value{
The `logger`type. If no corresponding logger-type is found,
  the result will be `NULL`.
}
\description{
Mapping the log-types from string to logger::<type>.
  E.g. the string "Info" will be mapped to `logger::INFO`.
}
