% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_round.R
\name{robust_round}
\alias{robust_round}
\title{Round numbers without problems.}
\usage{
robust_round(x, digits = 2, thousands_separator = "", decimal_separator = ".")
}
\arguments{
\item{x}{(numeric) The numeric input vector to round.}

\item{digits}{(int, optional) The number of digits after the decimal
separator to round to.}

\item{thousands_separator}{(string, optional) Used as mark between every 3
decimals before the decimal point.}

\item{decimal_separator}{(string, optional) The character to be used to
indicate the numeric decimal point.}
}
\value{
Rounded numbers as string.
}
\description{
Round numbers without problems.
}
\examples{
{
  robust_round(c(1.234567, 987123.987654321))
  #> [1] "1.23"   "987.99"
}
}
